# services/auth/api/docs/swagger.yaml
definitions:
  github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CompanyPrimaryKey:
    properties:
      id:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RegisterCompanyRequest:
    properties:
      fare_id:
        type: string
      googleToken:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RegisterCompanyRequest_GoogleToken'
      name:
        type: string
      user_info:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RegisterCompanyRequest_RegisterUserInfo'
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RegisterCompanyRequest_GoogleToken:
    properties:
      access_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RegisterCompanyRequest_RegisterUserInfo:
    properties:
      email:
        type: string
      expires_at:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateCompanyRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_company_service.GetListConfiguredResourceEnvironmentRes:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_company_service.GetListConfiguredResourceEnvironmentResResourceEnvironment'
        type: array
    type: object
  ? github_com_Ucode-io_ucode-opensource_services_auth_genproto_company_service.GetListConfiguredResourceEnvironmentResResourceEnvironment
  : properties:
      description:
        type: string
      display_color:
        type: string
      environment_id:
        type: string
      id:
        type: string
      is_configured:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      resource_id:
        type: string
      resource_type:
        type: integer
      service_type:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://u-code.io
paths:
...
  /company:
    post:
      consumes:
      - application/json
      description: Register Company
      operationId: create_company
      parameters:
      - description: RegisterCompanyRequestBody
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RegisterCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Company data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CompanyPrimaryKey'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Register Company
      tags:
      - Company
    put:
      consumes:
      - application/json
      description: Update Company
      operationId: update_company
      parameters:
      - description: UpdateCompanyRequestBody
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Company
      tags:
      - Company
  /company/{company-id}:
    delete:
      consumes:
      - application/json
      description: Get Company
      operationId: remove_company
      parameters:
      - description: company-id
        in: path
        name: company-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Remove Company
      tags:
      - Company
...
  /v2/resource-environment:
    get:
      consumes:
      - application/json
      description: Get Environments
      operationId: get_environments
      parameters:
      - description: project-id
        in: query
        name: project-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Environment data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_company_service.GetListConfiguredResourceEnvironmentRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Environments
      tags:
      - V2_Environment
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"